Add database and page creation
Implement POST, PATCH and GET for pages and databases.

Update to API version 2021-08-16.

Add a create_database.py example which creates a database, a page within it and retrieves again before deleting.

Run pre-commit autoupdate.

Run examples as part of tests (test_examples.py) if the required environment variables are set.

Continue separating the create, update and get types for pages and databases.